authentication middleware login and register with logout functionality and flash message created

+++ App/controllers/ListingController.php
@@ -2,8 +2,9 @@
 
 namespace App\Controllers;
 use Framework\Database;
+use Framework\Session;
 use Framework\Validation;
-
+use Framework\Authorization;
 class ListingController {
 
     protected $db;
@@ -15,7 +16,7 @@ class ListingController {
 
     public function index()
     {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
         
@@ -32,7 +33,7 @@ class ListingController {
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -69,6 +70,8 @@ class ListingController {
             $query = "INSERT INTO listings ({$fields}) VALUES ({$values})";
 
             $this->db->query($query, $newListingData);
+            Session::setFlashMessage('success_message', 'Listing Created Successfully');
+
             redirect('/listings');
         }
     }
@@ -105,9 +108,15 @@ class ListingController {
             ErrorController::notFound('Listing Not Found');
         }
 
+        if(!Authorization::isQwner($listing->user_id))
+        {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+           return redirect('/listings/' . $listing->id);
+        }
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
 
-        $_SESSION['success_message'] = 'Listing Deleted Successfully';
+        Session::setFlashMessage('success_message', 'Listing Deleted Successfully');
+
         redirect('/listings');
     }
 
@@ -126,7 +135,11 @@ class ListingController {
             ErrorController::notFound('Listing not found');
             return;
         }
-
+        if(!Authorization::isQwner($listing->user_id))
+        {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+           return redirect('/listings/' . $listing->id);
+        }
         loadView('listings/edit', [
             'listing' => $listing
         ]);
@@ -147,7 +160,11 @@ class ListingController {
             ErrorController::notFound('Listing not found');
             return;
         }
-
+        if(!Authorization::isQwner($listing->user_id))
+        {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+           return redirect('/listings/' . $listing->id);
+        }
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
    
         $updateValues = [];
@@ -179,8 +196,7 @@ class ListingController {
             $updateQuery = "UPDATE listings SET $updateFields WHERE id = :id";
             $updateValues['id'] = $id;
             $this->db->query($updateQuery, $updateValues);
-
-            $_SESSION['success_message'] = 'Listing Updated Successfully';
+            Session::setFlashMessage('success_message', 'Listing Updated Successfully');
             redirect('/listings/'. $id);
         }
     }
